% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_check.R
\name{df_check}
\alias{df_check}
\title{df_check}
\usage{
df_check(df)
}
\arguments{
\item{df}{An object the user intends to pass through manipulation functions. It is expected to be a data.frame.}
}
\value{
Invisibly returns TRUE if the input is a valid data.frame.
}
\description{
This function makes sure that the input into the function is a data.frame. If not, it stops execution and gives an informative error message.
}
\examples{
\dontrun{
check_dataframe(mtcars)   # passes
check_dataframe(1:10)     # throws an error
}
}
\keyword{internal}
