% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_rows.R
\name{filter_rows}
\alias{filter_rows}
\title{filter_rows}
\usage{
filter_rows(df, condition)
}
\arguments{
\item{df}{The input object given by the user. It is expected to be a data.frame.}

\item{condition}{A logical expression evaluated in the context of \code{df}. Logical conditions can be strung together by & or |.}
}
\value{
A filtered data.frame containing only the rows where the condition is TRUE.
}
\description{
This function returns rows of a data.frame that satisfy a logical condition. It is a simplified version of dplyr::filter() that is written in base R.
}
\examples{
#Keep only cars with mpg > 20
filter_rows(mtcars, mpg > 20)

#Keep only cars with mpg > 20 and cyl < 6
filter_rows(mtcars, mpg > 20 & cyl < 6)

}
