% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello World Function}
\usage{
hello(x)
}
\arguments{
\item{x}{Any input given}
}
\value{
NULL, prints the statement 'Hello, world!'
}
\description{
This function returns the phrase 'Hello, world!'.
}
\examples{
hello(x)
}
