% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{print_function}
\alias{print_function}
\title{Print Function}
\usage{
print_function(x)
}
\arguments{
\item{x}{Any parameter given}
}
\value{
NULL, prints the statement 'print function'
}
\description{
This function returns the phrase "print function".
}
\examples{
print_function(1)
}
