% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_columns.R
\name{select_columns}
\alias{select_columns}
\title{select_columns}
\usage{
select_columns(df, columns = NULL, exclude = NULL)
}
\arguments{
\item{df}{The input object given by the user. It is expected to be a data.frame.}

\item{columns}{A character or numeric vector specifying which columns to include (can be column names or numeric indices). Cannot be used together with \code{exclude}. Columns must exist in \code{df}.}

\item{exclude}{A character or numeric vector specifying which columns to exclude (can be column names or numeric indices). Cannot be used together with \code{columns}. Columns must exist in \code{df}.}
}
\value{
A filtered data.frame that contains only the specified columns, with the order of the columns being equivalent to the original data.frame (besides dropped columns).
}
\description{
This function returns a data.frame containing only the specified columns. Users can provide \code{columns} to include specific columns, or \code{exclude} to remove specific columns. Specified columns can be provided as names (character vector) or numeric indices. The function is a simplified version of dplyr::select() that is written in base R.
}
\examples{
# Select columns by name
select_columns(mtcars, columns = c("mpg", "hp"))

# Select columns by index
select_columns(mtcars, columns = c(1,5))

# Exclude columns by name
select_columns(mtcars, exclude = c("mpg", "hp"))

#Exclude columns by index
select_columns(mtcars, exclude = c(1,2))

# Attempting to select a non-existent column triggers an error
\dontrun{
select_columns(mtcars, columns = c("mpg", "blue_hair"))
}

}
